backend/ai_agents/rag/requirements.txt
# RAG Module Requirements

# Core dependencies
pydantic>=2.0.0
numpy>=1.20.0
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
langchain>=0.1.0
langchain-core>=0.1.0

# Embedding model dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Async support
aiohttp>=3.8.0

# Logging and utilities
tqdm>=4.65.0